+++ modules/data_source.py
@@ -8,25 +8,17 @@ import itertools
 from config.config import CHUNK_SIZE
 import queue
 
-def allocation_simulate(asset_allocation, assets, asset_revenue_per_year):
-    portfolio = Portfolio(assets=assets, weights=asset_allocation)
-    portfolio.simulate(asset_revenue_per_year)
-    return portfolio
-
-def allocation_simulate_serialize(asset_allocation, assets, asset_revenue_per_year):
-    portfolio = Portfolio(assets=assets, weights=asset_allocation)
-    portfolio.simulate(asset_revenue_per_year)
-    return portfolio.serialize()
 
 def allocation_simulate_serialize_slice_to_sink(slice_idx, slice_size, assets, percentage_step, asset_revenue_per_year, sink):
     portfolios_sent = 0
     with concurrent.futures.ThreadPoolExecutor() as thread_executor:
         possible_allocations_gen = all_possible_allocations(assets, percentage_step)
-        slice_gen = itertools.islice(possible_allocations_gen, slice_idx * slice_size, (slice_idx + 1) * slice_size)
-        simulate_func = partial(allocation_simulate_serialize, assets=assets, asset_revenue_per_year=asset_revenue_per_year)
-        simulated_portfolios = map(simulate_func, slice_gen)
+        gen_slice_allocations = itertools.islice(possible_allocations_gen, slice_idx * slice_size, (slice_idx + 1) * slice_size)
+        gen_portfolios = map(partial(Portfolio, assets=assets), gen_slice_allocations)
+        gen_simulateds = map(partial(Portfolio.simulated, asset_revenue_per_year=asset_revenue_per_year), gen_portfolios)
+        gen_serializations = map(Portfolio.serialize, gen_simulateds)
         send_task = None
-        for batch in itertools.batched(simulated_portfolios, CHUNK_SIZE):
+        for batch in itertools.batched(gen_serializations, CHUNK_SIZE):
             if send_task is not None:
                 send_task.result()
             send_task = thread_executor.submit(sink.send_bytes, b''.join(batch))
@@ -62,45 +54,6 @@ def all_possible_allocations(assets: list, step: int):
         asset_idx = 0, asset_idx_max = len(assets) - 1,
         allocation = [0] * len(assets),
         allocation_sum = 0)
-
-def all_possible_allocations_to_queue(assets: list, step: int, sink: queue.Queue):
-    for possible_allocation in all_possible_allocations(assets, step):
-        sink.put(possible_allocation)
-    return
-
-def queue_batched(future, source: queue.Queue, batch_size: int):
-    chunk = []
-    while not source.empty() and not future.done():
-        chunk.append(source.get())
-        if len(chunk) >= batch_size:
-            yield chunk
-            chunk = []
-    if len(chunk) > 0:
-        yield chunk
-
-    # # allocation_limit = 10*1000*1000
-    # # allocation_limit = 1000*1000
-    # # allocation_limit = 200*1000
-    # # allocation_limit = 10*1000
-
-    # time_start = time.time()
-
-    # # possible_allocations_gen = itertools.islice(all_possible_allocations(assets, percentage_step), allocation_limit)
-    # possible_allocations_gen = all_possible_allocations(assets, percentage_step)
-    # simulate_func = partial(allocation_simulate_serialize, assets=assets, asset_revenue_per_year=asset_revenue_per_year)
-    # total_portfolios = 0
-    # send_request = None
-    # for possible_allocation_batch in queue_batched(allocations_future, allocations_queue, CHUNK_SIZE):
-    #     simulated_portfolios_batch = process_pool.map(simulate_func, possible_allocation_batch)
-    #     total_portfolios += len(simulated_portfolios_batch)
-    #     bytes = b''.join(simulated_portfolios_batch)
-    #     # if send_request is not None:
-    #         # send_request.result()
-    #     # send_request = thread_pool.submit(sink.send_bytes, bytes)
-    # # send_request.result()
-    # thread_pool.shutdown()
-    # sink.send(data_types.DataStreamFinished())
-    # logger.info(f'Simulated {total_portfolios} portfolios, rate: {int(total_portfolios / (time.time() - time_start))}/s')
 
 
 def read_capitalgain_csv_data(filename):
